+++ tamr_unify_client/dataset/resource.py
@@ -144,8 +144,8 @@ class Dataset(BaseResource):
         :returns: The refresh operation.
         :rtype: :class:`~tamr_unify_client.operation.Operation`
         """
-        op_json = self.client.post(self.api_path + ":refresh").successful().json()
-        op = Operation.from_json(self.client, op_json)
+        response = self.client.post(self.api_path + ":refresh").successful()
+        op = Operation.from_response(self.client, response)
         return op.apply_options(**options)
 
     def profile(self):
@@ -174,10 +174,8 @@ class Dataset(BaseResource):
         :return: The operation to create the profile.
         :rtype: :class:`~tamr_unify_client.operation.Operation`
         """
-        op_json = (
-            self.client.post(self.api_path + "/profile:refresh").successful().json()
-        )
-        op = Operation.from_json(self.client, op_json)
+        response = self.client.post(self.api_path + "/profile:refresh").successful()
+        op = Operation.from_response(self.client, response)
         return op.apply_options(**options)
 
     def records(self):