+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 0.11.0-dev
+## 0.12.0-dev
+
+## 0.11.0
   **BETA**
 
   Important: Do not use BETA features for production workflows.